feat: Extracting build activity body
